Coordinates from int to double

+++ include/rrt_planner/node_2d.h
@@ -62,9 +62,9 @@ class Node2D {
    */
   struct Coordinates {
     Coordinates() {}
-    Coordinates(const int& x_in, const int& y_in) : x(x_in), y(y_in) {}
+    Coordinates(const double& x_in, const double& y_in) : x(x_in), y(y_in) {}
 
-    int x{0}, y{0};
+    double x{0}, y{0};
   };
   typedef std::vector<Coordinates> CoordinatesVector;
